Calculate between-census annual growth

Requires a data frame where separate measurements of the same tree appear in different rows and returns a data frame containing annual growth between each pair of consecutive measurements for each tree.

detailed_growth(data)

Arguments

data
A data frame containing repeated dbh measurements of trees. Must have a column named "tree_id" that contains unique tree ID values, a column named "year" containing the year of the measurement, and a column named "dbh" containing the dbh measurements. See built-in dataset tree for an example.

Value

A data frame containing between-census annual growth of each tree:

tree_id

unique identification code of the tree

start_year

first year of between-census period

begin_size

diameter at breast height at beginning of between-census period, in cm

end_year

last year of between-census period

end_size

diameter at breast height at end of between-census period, in cm

annual_growth

annual growth rate as average yearly increase in diameter at breast height, in cm/year

annual_bai

annual basal area increment as average yearly increase in area at breast height, in cm^2/year

size_corr_growth

square root of annual_growth divided by begin_size to give somewhat normally distributed growth rates

size_corr_growth_basal

square root of annual_bai divided by begin_size to give somewhat normally distributed growth rates

...

any other columns that appeared in data

Examples

detailed_growth(tree)

#> [1] "Warning: 2913 cases of negative annual growth"

#> # A tibble: 50,959 x 12
#> # Groups:   tree_id [8,721]
#>    tree_id       stand_id   tag species start_year begin_size end_year end_size
#>    <chr>         <chr>    <int> <chr>        <int>      <dbl>    <int>    <dbl>
#>  1 AB08000100001 AB08      4001 TSHE          1978       48.2     1984     48.2
#>  2 AB08000100001 AB08      4001 TSHE          1984       48.2     1990     48.6
#>  3 AB08000100001 AB08      4001 TSHE          1990       48.6     1995     49  
#>  4 AB08000100001 AB08      4001 TSHE          1995       49       2002     49.6
#>  5 AB08000100001 AB08      4001 TSHE          2002       49.6     2008     50.3
#>  6 AB08000100001 AB08      4001 TSHE          2008       50.3     2012     50.4
#>  7 AB08000100001 AB08      4001 TSHE          2012       50.4     2017     50.6
#>  8 AB08000100002 AB08      4002 TSHE          1978       76.3     1984     76.8
#>  9 AB08000100002 AB08      4002 TSHE          1984       76.8     1990     77.3
#> 10 AB08000100002 AB08      4002 TSHE          1990       77.3     1995     78.7
#> # ... with 50,949 more rows, and 4 more variables: annual_growth <dbl>,
#> #   annual_bai <dbl>, size_corr_growth <dbl>, size_corr_growth_basal <dbl>