Calculate annual growth

Requires a data frame where separate measurements of the same tree appear in different rows and returns a data frame with a single row for each tree containing annual growth measurements.

growth_summary(data)

Arguments

data
A dataframe containing repeated measurements of growth. Must have a column named "tree_id" that contains unique tree ID values, a column named "year" containing the year of the measurement, and a column named "dbh" containing the dbh measurements.See built-in dataset tree for an example.

Value

A dataframe containing growth rate measurements for each tree:

tree_id

unique identification code of the tree

stand_id

name of stand in which tree is located - only appears if input dataframe contains a stand_id column

species

species identity of tree as a four letter code

first_record

year of first size measurement of the tree

last_record

year of last size measurement of the tree

begin_size

first size measurement as diameter at breast height, in cm

final_size

last size measurement as diameter at breast height, in cm

mean_size

average size across measurements as diameter at breast height, in cm

midpoint_size

average of maximum and minimum size across measurements as diameter at breast height, in cm

annual_growth

annual growth rate as average yearly increase in diameter at breast height, in cm/year

annual_bai

annual basal area increment as average yearly increase in area at breast height, in cm^2/year

size_corr_growth

square root of annual_growth divided by begin_size to give somewhat normally distributed growth rates

Examples

growth_summary(tree)

#> [1] "Warning: 139 trees exhibited negative annual growth"

#> # A tibble: 9,155 x 12
#>    tree_id       stand_id species first_record last_record begin_size final_size
#>    <chr>         <chr>    <chr>          <int>       <int>      <dbl>      <dbl>
#>  1 AB08000100001 AB08     TSHE            1978        2017       48.2       50.6
#>  2 AB08000100002 AB08     TSHE            1978        2017       76.3       81.6
#>  3 AB08000100003 AB08     TSHE            1978        2017       15.7       20.2
#>  4 AB08000100004 AB08     TSHE            1978        2017       50.5       54  
#>  5 AB08000100005 AB08     TSHE            1978        2017       18.5       24.5
#>  6 AB08000100006 AB08     ABAM            1978        2017       25.2       30.4
#>  7 AB08000100007 AB08     THPL            1978        2017       47.4       58.2
#>  8 AB08000100008 AB08     TSHE            1978        2017       20.3       26.9
#>  9 AB08000100009 AB08     THPL            1978        2017       48         56.5
#> 10 AB08000100010 AB08     PSME            1978        2017       79.6       83  
#> # ... with 9,145 more rows, and 5 more variables: mean_size <dbl>,
#> #   midpoint_size <dbl>, annual_growth <dbl>, annual_bai <dbl>,
#> #   size_corr_growth <dbl>