Summarize neighborhoods

Calculates species-specific densities and species richness for each neighborhood in the provided data frame. Densities can be provided as area covered in the neighborhood, area covered per hectare, or the sum of horizontal angles.

neighborhood_summary(
  neighbors,
  id_column,
  radius,
  densities = "raw",
  neighbors_incl = "all",
  edge_correction = F,
  x_limit = NULL,
  y_limit = NULL
)

Arguments

neighbors
Data frame of neighborhoods outputted by neighborhoods function.

id_column
Name of column in neighbors containing site names as a string.

radius
Numeric vector describing neighborhood radius in meters.

densities
Character specifying the type of density measurements to calculate - raw (m^2 per hectare), proportional (as proportions of overall tree density), angular (angular size of trees).

neighbors_incl
Character specifying whether all neighbors (neighbors_incl = "all") or only neighbors larger than the focal tree (neighbors_incl = "larger_only") should be included in diversity and density calculations.

edge_correction
Boolean indicating whether edge correction should be used.

x_limit
maximum possible x-coordinate in the stand (only required if edge_correction = T).

y_limit
maximum possible y-coordinate in the stand (only required if edge_correction = T).

Value

Data frame containing a summary for each neighborhood in neighbors.

Details

This function contains an optional argument for handling edge effects. When specified, the density and species richness values for neighborhoods that overlap the stand boundary will be multiplied according to the fraction of their neighborhood that is missing from the stand. Species richness values are rounded to the nearest whole number.

Examples

# Create a neighborhoods object
nbhds <- neighborhoods(mapping, stands = "AB08", radius = 10)

# Summarize neighborhoods using angular densities
nbhd_summ <- neighborhood_summary(nbhds, id_column = "tree_id", radius = 10,
                                  densities = "angular")

# Using raw densities with edge correction
nbhd_summ <- neighborhood_summary(nbhds, id_column = "tree_id", radius = 10,
                                  densities = "raw", edge_correction = T,
                                  x_limit = 100, y_limit = 100)